requirements.txt
asgiref==3.11.0
charset-normalizer==3.4.4
Django==5.1
pillow==12.1.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
reportlab==4.4.6
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
watchdog==6.0.0
gspread==6.0.2
google-auth==2.29.0
requests==2.31.0
